+++ lib/commanded/aggregates/supervisor.ex
@@ -10,10 +10,6 @@ defmodule Commanded.Aggregates.Supervisor do
   alias Commanded.Aggregates.Aggregate
   alias Commanded.Registration
 
-  # def child_spec(arg) do
-  #   Registration.supervisor_child_spec(__MODULE__, arg)
-  # end
-
   def start_link(args) do
     application = Keyword.fetch!(args, :application)
     name = Module.concat([application, __MODULE__])
@@ -35,7 +31,7 @@ defmodule Commanded.Aggregates.Supervisor do
         inspect(aggregate_uuid)
     end)
 
-    module_name = Module.concat([application, __MODULE__])
+    supervisor_name = Module.concat([application, __MODULE__])
     aggregate_name = Aggregate.name(application, aggregate_module, aggregate_uuid)
 
     args = [
@@ -44,7 +40,7 @@ defmodule Commanded.Aggregates.Supervisor do
       aggregate_uuid: aggregate_uuid
     ]
 
-    case Registration.start_child(application, aggregate_name, module_name, {Aggregate, args}) do
+    case Registration.start_child(application, aggregate_name, supervisor_name, {Aggregate, args}) do
       {:ok, _pid} ->
         {:ok, aggregate_uuid}