+++ test/application/application_test.exs
@@ -2,8 +2,9 @@ defmodule Commanded.ApplicationTest do
   use ExUnit.Case
 
   alias Commanded.Application
-  alias Commanded.EventStore.EventData
   alias Commanded.ExampleApplication
+  alias Commanded.RuntimeConfiguredApplication
+  alias Commanded.UnconfiguredApplication
 
   describe "a Commanded application" do
     setup do
@@ -12,56 +13,60 @@ defmodule Commanded.ApplicationTest do
       [pid: pid]
     end
 
-    defmodule Event do
-      @derive Jason.Encoder
-      defstruct [:name]
-    end
-
     test "should not allow an application to be started more than once", %{pid: pid} do
       assert {:error, {:already_started, ^pid}} = ExampleApplication.start_link()
     end
 
     test "should expose an event store adapter" do
-      assert Application.event_store_adapter(ExampleApplication) == ExampleApplication.EventStore
-    end
-
-    test "should define an event store adapter" do
-      assert_implements(ExampleApplication.EventStore, Commanded.EventStore.Adapter)
-
-      events = [
-        %EventData{
-          correlation_id: UUID.uuid4(),
-          causation_id: UUID.uuid4(),
-          event_type: "#{__MODULE__}.Event",
-          data: %Event{name: "1"},
-          metadata: %{"metadata" => "value"}
-        }
-      ]
-
-      assert :ok = ExampleApplication.EventStore.append_to_stream("1", 0, events)
+      assert Application.event_store_adapter(ExampleApplication) ==
+               {Commanded.EventStore.Adapters.InMemory,
+                %{name: Commanded.ExampleApplication.EventStore}}
     end
 
     test "should expose a pubsub adapter" do
-      assert Application.pubsub_adapter(ExampleApplication) == ExampleApplication.PubSub
-    end
-
-    test "should define a pubsub adapter" do
-      assert_implements(ExampleApplication.PubSub, Commanded.PubSub.Adapter)
+      assert Application.pubsub_adapter(ExampleApplication) ==
+               {Commanded.PubSub.LocalPubSub,
+                %{
+                  pubsub_name: Commanded.ExampleApplication.LocalPubSub,
+                  tracker_name: Commanded.ExampleApplication.LocalPubSub.Tracker
+                }}
     end
 
     test "should expose a registry adapter" do
-      assert Application.registry_adapter(ExampleApplication) == ExampleApplication.Registration
+      assert Application.registry_adapter(ExampleApplication) ==
+               {Commanded.Registration.LocalRegistry,
+                %{registry_name: Commanded.ExampleApplication.LocalRegistry}}
     end
 
-    test "should define a registration adapter" do
-      assert_implements(ExampleApplication.Registration, Commanded.Registration.Adapter)
+    test "should allow application config to be provided by `init/1` callback function" do
+      start_supervised!(RuntimeConfiguredApplication)
+
+      assert Application.event_store_adapter(RuntimeConfiguredApplication) ==
+               {Commanded.EventStore.Adapters.InMemory,
+                %{name: Commanded.RuntimeConfiguredApplication.EventStore}}
+
+      assert Application.pubsub_adapter(RuntimeConfiguredApplication) ==
+               {Commanded.PubSub.PhoenixPubSub,
+                %{
+                  pubsub_name: Commanded.RuntimeConfiguredApplication.PhoenixPubSub,
+                  tracker_name: Commanded.RuntimeConfiguredApplication.PhoenixPubSub.Tracker
+                }}
+
+      assert Application.registry_adapter(RuntimeConfiguredApplication) ==
+               {Commanded.Registration.LocalRegistry,
+                %{registry_name: Commanded.RuntimeConfiguredApplication.LocalRegistry}}
     end
-  end
 
-  # Returns `true` if module implements behaviour.
-  defp assert_implements(module, behaviour) do
-    all = Keyword.take(module.__info__(:attributes), [:behaviour])
+    test "should fail to start unconfigured application" do
+      Process.flag(:trap_exit, true)
 
-    assert [behaviour] in Keyword.values(all)
+      UnconfiguredApplication.start_link()
+
+      assert_receive {:EXIT, _pid,
+                      {%ArgumentError{
+                         message:
+                           "missing :event_store config for application Commanded.UnconfiguredApplication"
+                       }, _}}
+    end
   end
 end